+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Unblock npm install -g + CI TS errors for v0.34.1
 ---
 
 ## [0.34.0] - 2026-04-08